refactor(watchhistory): separate useWatchHistory hook and add progressive save interval
